Fine-Tunning/requirements.txt
# Core ML/AI Libraries
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
peft>=0.6.0
bitsandbytes>=0.41.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
datasets>=2.14.0

# Evaluation Metrics
rouge-score>=0.1.2
sacrebleu>=2.3.0
nltk>=3.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Development
jupyter>=1.0.0
ipywidgets>=8.0.0
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0
